Move slice utils to service layer
